more comments on #519
