+++ ext/PolynomialsFFTWExt.jl
@@ -1,22 +1,9 @@
 module PolynomialsFFTWExt
 
 using Polynomials
-import Polynomials: MutableDensePolynomial, StandardBasis
+import Polynomials: MutableDensePolynomial, StandardBasis, Pad
 import FFTW
 import FFTW: fft, ifft
-
-struct Pad{T} <: AbstractVector{T}
-    a::Vector{T}
-    n::Int
-end
-Base.length(a::Pad) = a.n
-Base.size(a::Pad) = (a.n,)
-function Base.getindex(a::Pad, i)
-    u = length(a.a)
-    i ≤ u && return a.a[i]
-    return zero(first(a.a))
-end
-
 function Polynomials.poly_multiplication_fft(p::P, q::Q) where {B <: StandardBasis,X,
                                                                 T <: AbstractFloat, P<:MutableDensePolynomial{B,T,X},
                                                                 S <: AbstractFloat, Q<:MutableDensePolynomial{B,S,X}}
